Work in progress paper cuts document
I've been working on this in Notes for some time but I'm just now beginning to
close in on a document of the quality required for sharing.

Rather than wait until the last moment and commit the Perfect Engagement
Document I have decided to record it now and give myself the ability to look
back in time.